Add key management and GPG signing
This commit builds out most of the required machinery to implement the
actual interesting workflows (signing RPMs, etc). There's still a great
deal of refactoring and polish to add, as well as more comprehensive
test coverage. However, since this is all new and pre-release, that will
be done in smaller follow-up PRs.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>