Add signing with OpenSSL
Support signing using RSA and ECDSA using OpenSSL. There's an API for
both pre-hashed content (where the client just sends us a hash) and
blobs (where we hash the content). In practice, it may be we only ever
want to use the pre-hashed APIs, so we should re-evaluate this once the
client side is fleshed out.

OpenSSL has a lot of caveats around signing with Ed25519. For now, let's
not support it since it's not currently used in Fedora anyway.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>