Bump Rust MSRV to 1.88 and enable 2024 edition
This project tracks the latest EL Rust version; with 10.1 and 9.7 this
is 1.88. This also allows us to switch to the 2024 edition.

The one incompatibility was that for the tests involving SoftHSM, we
have to write the configuration file's location to an environment
variable. We specifically switched to nextest to make this safe, and
with the 2024 edition the `set_env` function is marked unsafe. We can
continue to use it, but the lint forbidding unsafe code needs to be
removed.

All other changes are due to rustfmt changes.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>