siguldry-pkcs11: implement a PKCS #11 module for siguldry
This implements a bare-minimum PKCS#11 module to use siguldry as the
backend. Notably, it is not a complete implementation of the PKCS#11
specification, as it supports only signing operations. At this time,
management operations such as key creation are not supported.

This module is also not likely to win any specification compliance
awards as some mechanisms, like the CKM_RSA_PKCS and CKM_ECDSA
mechanisms limit their accepted inputs to 32 and 64 bytes (presumed to
be SHA256 and SHA512 respectively) and, in the case of CKM_RSA_PKCS,
DER-encoded DigestInfo structures that include the digest algorithm and
digest itself.

OpenPGP keys on the Siguldry server are exposed using the format Sequoia
expects with its (experimental) PKCS11 support.

The final oddity with this module is that it communicates with the
Siguldry server using a local Unix socket to proxy requests. This is for
two reasons.

The first is that the module is often used via OpenSSL's pkcs11
provider, and because Siguldry also uses OpenSSL with the default
context, this results in some surprising interactions.

The second reason is that if we work over a Unix socket, it's possible
to provide the socket inside an isolated build environment and allow the
build to install this module and sign content during the build. The
obvious use-case for that is Secure Boot signatures for UEFI
applications Fedora ships since that's how pesign-daemon works. It also
works with any tool that understands pkcs11, so the same approach could
be used for any artifact that needs to sign something that's included in
an RPM.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>